requirements.txt
aws-cdk-lib==2.141.0
constructs>=10.0.0,<11.0.0
cdk-dynamo-table-view==0.2.487